changed IntegerLiteral's symbol from Integer to Long; many tests fail now

+++ plugins/org.eclipse.ocl.ecore/src/org/eclipse/ocl/ecore/impl/IntegerLiteralExpImpl.java
@@ -55,7 +55,7 @@ public class IntegerLiteralExpImpl
 	 * @generated
 	 * @ordered
 	 */
-	protected static final Integer INTEGER_SYMBOL_EDEFAULT = null;
+	protected static final Long INTEGER_SYMBOL_EDEFAULT = null;
 
 	/**
 	 * The cached value of the '{@link #getIntegerSymbol() <em>Integer Symbol</em>}' attribute.
@@ -65,7 +65,7 @@ public class IntegerLiteralExpImpl
 	 * @generated
 	 * @ordered
 	 */
-	protected Integer integerSymbol = INTEGER_SYMBOL_EDEFAULT;
+	protected Long integerSymbol = INTEGER_SYMBOL_EDEFAULT;
 
 	/**
 	 * <!-- begin-user-doc -->
@@ -90,8 +90,9 @@ public class IntegerLiteralExpImpl
 	 * <!-- begin-user-doc -->
 	 * <!-- end-user-doc -->
 	 * @generated
+	 * @since 3.2
 	 */
-	public Integer getIntegerSymbol() {
+	public Long getIntegerSymbol() {
 		return integerSymbol;
 	}
 
@@ -99,9 +100,10 @@ public class IntegerLiteralExpImpl
 	 * <!-- begin-user-doc -->
 	 * <!-- end-user-doc -->
 	 * @generated
+	 * @since 3.2
 	 */
-	public void setIntegerSymbol(Integer newIntegerSymbol) {
-		Integer oldIntegerSymbol = integerSymbol;
+	public void setIntegerSymbol(Long newIntegerSymbol) {
+		Long oldIntegerSymbol = integerSymbol;
 		integerSymbol = newIntegerSymbol;
 		if (eNotificationRequired())
 			eNotify(new ENotificationImpl(this, Notification.SET,
@@ -143,7 +145,7 @@ public class IntegerLiteralExpImpl
 	public void eSet(int featureID, Object newValue) {
 		switch (featureID) {
 			case EcorePackage.INTEGER_LITERAL_EXP__INTEGER_SYMBOL :
-				setIntegerSymbol((Integer) newValue);
+				setIntegerSymbol((Long) newValue);
 				return;
 		}
 		super.eSet(featureID, newValue);
