+++ plugins/org.eclipse.ocl/src/org/eclipse/ocl/cst/impl/IntegerLiteralExpCSImpl.java
@@ -48,7 +48,7 @@ public class IntegerLiteralExpCSImpl
 	 * @generated
 	 * @ordered
 	 */
-	protected static final Integer INTEGER_SYMBOL_EDEFAULT = null;
+	protected static final Long INTEGER_SYMBOL_EDEFAULT = null;
 
 	/**
 	 * The cached value of the '{@link #getIntegerSymbol() <em>Integer Symbol</em>}' attribute.
@@ -58,7 +58,7 @@ public class IntegerLiteralExpCSImpl
 	 * @generated
 	 * @ordered
 	 */
-	protected Integer integerSymbol = INTEGER_SYMBOL_EDEFAULT;
+	protected Long integerSymbol = INTEGER_SYMBOL_EDEFAULT;
 
 	/**
 	 * <!-- begin-user-doc -->
@@ -84,7 +84,7 @@ public class IntegerLiteralExpCSImpl
 	 * <!-- end-user-doc -->
 	 * @generated
 	 */
-	public Integer getIntegerSymbol() {
+	public Long getIntegerSymbol() {
 		return integerSymbol;
 	}
 
@@ -93,8 +93,8 @@ public class IntegerLiteralExpCSImpl
 	 * <!-- end-user-doc -->
 	 * @generated
 	 */
-	public void setIntegerSymbol(Integer newIntegerSymbol) {
-		Integer oldIntegerSymbol = integerSymbol;
+	public void setIntegerSymbol(Long newIntegerSymbol) {
+		Long oldIntegerSymbol = integerSymbol;
 		integerSymbol = newIntegerSymbol;
 		if (eNotificationRequired())
 			eNotify(new ENotificationImpl(this, Notification.SET,
@@ -125,7 +125,7 @@ public class IntegerLiteralExpCSImpl
 	public void eSet(int featureID, Object newValue) {
 		switch (featureID) {
 			case CSTPackage.INTEGER_LITERAL_EXP_CS__INTEGER_SYMBOL :
-				setIntegerSymbol((Integer) newValue);
+				setIntegerSymbol((Long) newValue);
 				return;
 		}
 		super.eSet(featureID, newValue);
