+++ plugins/org.eclipse.ocl/src/org/eclipse/ocl/expressions/impl/IntegerLiteralExpImpl.java
@@ -52,7 +52,7 @@ public class IntegerLiteralExpImpl<C>
 	 * @generated
 	 * @ordered
 	 */
-	protected static final Integer INTEGER_SYMBOL_EDEFAULT = null;
+	protected static final Long INTEGER_SYMBOL_EDEFAULT = null;
 
 	/**
 	 * The cached value of the '{@link #getIntegerSymbol() <em>Integer Symbol</em>}' attribute.
@@ -62,7 +62,7 @@ public class IntegerLiteralExpImpl<C>
 	 * @generated
 	 * @ordered
 	 */
-	protected Integer integerSymbol = INTEGER_SYMBOL_EDEFAULT;
+	protected Long integerSymbol = INTEGER_SYMBOL_EDEFAULT;
 
 	/**
 	 * <!-- begin-user-doc -->
@@ -88,7 +88,7 @@ public class IntegerLiteralExpImpl<C>
 	 * <!-- end-user-doc -->
 	 * @generated
 	 */
-	public Integer getIntegerSymbol() {
+	public Long getIntegerSymbol() {
 		return integerSymbol;
 	}
 
@@ -97,8 +97,8 @@ public class IntegerLiteralExpImpl<C>
 	 * <!-- end-user-doc -->
 	 * @generated
 	 */
-	public void setIntegerSymbol(Integer newIntegerSymbol) {
-		Integer oldIntegerSymbol = integerSymbol;
+	public void setIntegerSymbol(Long newIntegerSymbol) {
+		Long oldIntegerSymbol = integerSymbol;
 		integerSymbol = newIntegerSymbol;
 		if (eNotificationRequired())
 			eNotify(new ENotificationImpl(this, Notification.SET,
@@ -140,7 +140,7 @@ public class IntegerLiteralExpImpl<C>
 	public void eSet(int featureID, Object newValue) {
 		switch (featureID) {
 			case ExpressionsPackage.INTEGER_LITERAL_EXP__INTEGER_SYMBOL :
-				setIntegerSymbol((Integer) newValue);
+				setIntegerSymbol((Long) newValue);
 				return;
 		}
 		super.eSet(featureID, newValue);
